+++ packages/vsync_lab_toolkit/lib/vsync_lab_toolkit.dart
@@ -1,6 +1,8 @@
 library;
 
-export 'src/frame_log_file_exporter.dart'
-    show FrameLogExporter, FrameLogFileExporter, FrameLogSaveResult;
+export 'src/frame_log_exporter.dart' show FrameLogExporter, FrameLogSaveResult;
+export 'src/frame_log_file_exporter.dart' show FrameLogFileExporter;
+export 'src/frame_metrics_recorder.dart' show FrameMetricsRecorder;
 export 'src/frame_metrics_snapshot.dart' show FrameMetricsSnapshot;
+export 'src/frame_sample.dart' show FrameSample;
 export 'src/frame_timing_monitor.dart' show FrameTimingMonitor;